+++ src/commands/selectBuildTarget.ts
@@ -1,6 +1,6 @@
 import { Commands } from '@commands/commandIds';
 import { withCommand } from '@commands/withCommand';
-import { promptAndSaveBuildTarget } from '@services/buildService';
+import { promptAndSaveBuildTarget } from '@services/buildTargetService';
 
 export function registerSelectBuildTarget() {
 	return withCommand(Commands.selectBuildTarget, async () => {